Test an entire range of PSFs against camera
This commit holds the code testing a defined 3D space of PSFs
against the camera's behavior. It selects the "best" convolution
and attempts precorrection. In this case, the "best" is determined
with the lowest MSE between the convolved image and the captured
image, with their histograms matched, weighed by the convolved
image's gradient.