Display BC images, test PSF using precorrected images
The BC (boundary conditions) during convolution is assumed to be
either periodic or reflexive, but the captured images are previously
surrounded by black. This update simulates the periodic or reflexive
boundary conditions for captured images so that the edges won't
differ from those of the convolved images.

The PSF is also tested by comparing captured and convolved
precorrected images, rather than the original. This proves to have
better results since an ideal convolution smooths out and cancels
most of the ripples. If the PSF used differs from the camera, the
captured image won't cancel the ripples as successfully, so we'll
get a greater error. For this reason, I've disabled the gradient
weighing, since the errors will occur away from the edges of the
perfectly convolved image.