Move a bunch of code around
I've added a set of files exclusively for camera calibration that
currently include code for aligning and converting images. I also
added a set of files for common operations such as reading, writing,
and displaying images. All of this code is no longer in main. The
project compiles, but I haven't tested any of it yet.